microprojeto/pyproject.toml
[project]
name = "projeto-flask-jolly"
version = "0.1.0"
description = "Projeto Flask com Gemini e SQLite"
requires-python = ">=3.11"

dependencies = [
    "flask==3.0.0",
    "flask_sqlalchemy==3.1.1",
    "google-genai==0.2.0",
    "python-dotenv==1.0.0"
]